+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 
 ### Changed
 
+- Tweaked how Terminal.app reports capabilities
+
 ### Fixed
 
 ## 0.1.4 - 2025-11-09